Refactor: #8 batch관련클래스 코틀린마이그레이션

+++ src/main/java/com/back/domain/member/memberGame/dto/MemberGameDto.java
@@ -9,7 +9,7 @@ public record MemberGameDto(
         double playtime,
         boolean isFavorite,
         StatusEnum status,
-        int gameId,
+        Long gameId,
         Long igdbId,
         String gameName,
         String coverImageId,
